+++ src/infrastructure/mod.rs
@@ -1,4 +1,4 @@
 pub mod console_progress_reporter;
 pub mod dialoguer_ui_selector;
+pub mod seeder;
 pub mod system_environment;
-pub mod system_seeder;