+++ src/lib.rs
@@ -1,27 +1,32 @@
 use anyhow::Result;
+use console::style;
 
 mod application;
 mod domain;
 mod infrastructure;
 mod interfaces;
 
+use application::use_cases::generate_feature::GenerateFeatureUseCase;
 use application::use_cases::new_project::NewProjectUseCase;
 use infrastructure::console_progress_reporter::ConsoleProgressReporter;
 use infrastructure::dialoguer_ui_selector::DialoguerUiSelector;
+use infrastructure::seeder::SystemSeeder;
 use infrastructure::system_environment::SystemEnvironment;
-use infrastructure::system_seeder::SystemSeeder;
 
 const BANNER: &str = r#"
-        _   __      _____               __
-       / | / /___ _/ ___/___  ___  ____/ /
-      /  |/ / __ `/\__ \/ _ \/ _ \/ __  /
-     / /|  / /_/ /___/ /  __/  __/ /_/ /
-    /_/ |_/\__, //____/\___/\___/\__,_/
-          /____/
+
+    ______           _        _     _ _
+    | ___ \         | |      | |   (_) |
+    | |_/ /_ __ ___ | |_ __ _| |    _| |__  _ __ ___
+    | ___ \ '__/ _ \| __/ _` | |   | | '_ \| '__/ _ \
+    | |_/ / | | (_) | || (_| | |___| | |_) | | |  __/
+    \____/|_|  \___/ \__\__,_\_____/_|_.__/|_|  \___|
+
+    BrotaLibre - Grow scalable frontends from solid foundations.
 "#;
 
 pub fn run() -> Result<()> {
-    println!("{}", BANNER);
+    println!("{}", style(BANNER).blue());
 
     let command = interfaces::cli::parse()?;
 
@@ -33,6 +38,13 @@ pub fn run() -> Result<()> {
             let reporter = ConsoleProgressReporter::default();
 
             let use_case = NewProjectUseCase::new(&env, &ui_selector, &seeder, &reporter);
+            use_case.execute(request)
+        }
+        interfaces::cli::AppCommand::GenerateFeature(request) => {
+            let seeder = SystemSeeder;
+            let reporter = ConsoleProgressReporter::default();
+
+            let use_case = GenerateFeatureUseCase::new(&seeder, &reporter);
             use_case.execute(request)
         }
     }