+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.2.2](https://github.com/eehcx/ngseed/compare/ngseed-v0.2.1...ngseed-v0.2.2) (2026-03-27)
+
+
+### Bug Fixes
+
+* **ci:** resolve template path for installed binary ([#10](https://github.com/eehcx/ngseed/issues/10)) ([5c71473](https://github.com/eehcx/ngseed/commit/5c714730cc4ccb30915f3b20cc328a417d789b8b))
+
+## [0.2.1](https://github.com/eehcx/ngseed/compare/ngseed-v0.2.0...ngseed-v0.2.1) (2026-03-27)
+
+
+### Bug Fixes
+
+* create tailwind.config.js and postcss.config.js manually instead of running npx tailwindcss init ([#8](https://github.com/eehcx/ngseed/issues/8)) ([078eec4](https://github.com/eehcx/ngseed/commit/078eec49f31be22187d4b99f89ab78c172105e3d))
+
+## [0.2.0](https://github.com/eehcx/ngseed/compare/ngseed-v0.1.2...ngseed-v0.2.0) (2026-03-27)
+
+
+### Features
+
+* add templates system and cdp architecture profile ([#6](https://github.com/eehcx/ngseed/issues/6)) ([654c1be](https://github.com/eehcx/ngseed/commit/654c1be3f6511831348a150097ec6080588c136e))
+
 ## [0.1.2](https://github.com/eehcx/ngseed/compare/ngseed-v0.1.1...ngseed-v0.1.2) (2026-03-26)